# Conflicts:
#	app/schemas/com.fongmi.android.tv.db.AppDatabase/25.json
#	app/schemas/com.fongmi.android.tv.db.AppDatabase/33.json
#	app/schemas/com.fongmi.android.tv.db.AppDatabase/35.json
#	app/src/leanback/java/com/fongmi/android/tv/Updater.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/BufferDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/ConfigDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/DescDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/DohDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/HistoryDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/LiveDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/ProxyDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/SiteDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/dialog/UaDialog.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/presenter/ChannelPresenter.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/presenter/EpisodePresenter.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/presenter/FlagPresenter.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/presenter/GroupPresenter.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/presenter/PartPresenter.java
#	app/src/leanback/java/com/fongmi/android/tv/ui/presenter/QuickPresenter.java
#	app/src/leanback/res/drawable/shape_chip_activated.xml
#	app/src/leanback/res/drawable/shape_epg_focused.xml
#	app/src/leanback/res/drawable/shape_subtitle.xml
#	app/src/leanback/res/layout/dialog_proxy.xml
#	app/src/main/assets/css/style.css
#	app/src/main/assets/css/ui.css
#	app/src/main/assets/js/script.js
#	app/src/main/java/com/fongmi/android/tv/Setting.java
#	app/src/main/java/com/fongmi/android/tv/api/LiveParser.java
#	app/src/main/java/com/fongmi/android/tv/api/config/VodConfig.java
#	app/src/main/java/com/fongmi/android/tv/api/config/WallConfig.java
#	app/src/main/java/com/fongmi/android/tv/api/loader/JarLoader.java
#	app/src/main/java/com/fongmi/android/tv/api/loader/JsLoader.java
#	app/src/main/java/com/fongmi/android/tv/api/loader/PyLoader.java
#	app/src/main/java/com/fongmi/android/tv/bean/Hot.java
#	app/src/main/java/com/fongmi/android/tv/bean/Page.java
#	app/src/main/java/com/fongmi/android/tv/bean/Part.java
#	app/src/main/java/com/fongmi/android/tv/bean/Suggest.java
#	app/src/main/java/com/fongmi/android/tv/event/ErrorEvent.java
#	app/src/main/java/com/fongmi/android/tv/event/PlayerEvent.java
#	app/src/main/java/com/fongmi/android/tv/event/ScanEvent.java
#	app/src/main/java/com/fongmi/android/tv/impl/FilterCallback.java
#	app/src/main/java/com/fongmi/android/tv/impl/SessionCallback.java
#	app/src/main/java/com/fongmi/android/tv/player/Players.java
#	app/src/main/java/com/fongmi/android/tv/player/exo/TrackNameProvider.java
#	app/src/main/java/com/fongmi/android/tv/service/DLNACastService.java
#	app/src/main/java/com/fongmi/android/tv/service/PlaybackService.java
#	app/src/main/java/com/fongmi/android/tv/ui/custom/dialog/BaseDialog.java
#	app/src/main/java/com/fongmi/android/tv/ui/dialog/BaseBottomSheetDialog.java
#	app/src/main/java/com/fongmi/android/tv/ui/dialog/BaseDialog.java
#	app/src/main/java/com/fongmi/android/tv/ui/dialog/TrackDialog.java
#	app/src/main/java/com/fongmi/android/tv/utils/OkGlideModule.java
#	app/src/main/java/com/fongmi/android/tv/utils/PauseExecutor.java
#	app/src/main/res/drawable-nodpi/ic_logo.png
#	app/src/main/res/drawable/ic_notify_next.xml
#	app/src/main/res/drawable/ic_notify_pause.xml
#	app/src/main/res/drawable/ic_notify_play.xml
#	app/src/main/res/drawable/ic_notify_prev.xml
#	app/src/mobile/java/com/fongmi/android/tv/Updater.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/activity/CollectActivity.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/activity/HomeActivity.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/custom/CustomKeyDownLive.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/custom/CustomKeyDownVod.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/BufferDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/CastDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/ConfigDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/EpisodeListDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/HistoryDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/LinkDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/LiveDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/ProxyDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/SiteDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/SyncDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/dialog/UaDialog.java
#	app/src/mobile/java/com/fongmi/android/tv/ui/holder/VodOneHolder.java
#	app/src/mobile/java/com/fongmi/android/tv/utils/ScanTask.java
#	app/src/mobile/res/drawable/ic_notify_next.xml
#	app/src/mobile/res/drawable/ic_notify_pause.xml
#	app/src/mobile/res/drawable/ic_notify_play.xml
#	app/src/mobile/res/drawable/ic_notify_prev.xml
#	app/src/mobile/res/drawable/ic_subtitle_down.xml
#	app/src/mobile/res/drawable/ic_subtitle_large.xml
#	app/src/mobile/res/drawable/ic_subtitle_small.xml
#	app/src/mobile/res/drawable/ic_subtitle_up.xml
#	app/src/mobile/res/layout/activity_collect.xml
#	app/src/mobile/res/layout/dialog_proxy.xml
#	catvod/src/main/java/com/github/catvod/net/OkProxySelector.java
#	catvod/src/main/java/com/github/catvod/utils/Github.java
#	chaquo/src/main/java/com/fongmi/chaquo/Loader.java
#	other/image/icon.png
#	other/image/logo-2.png
#	other/sample/live/offline.json
#	other/sample/live/online.json
#	other/sample/vod/offline.json
#	other/sample/vod/online.json
#	quickjs/src/main/java/com/fongmi/quickjs/utils/Async.java

+++ app/src/mobile/java/com/fongmi/android/tv/ui/adapter/ConfigAdapter.java
@@ -1,6 +1,7 @@
 package com.fongmi.android.tv.ui.adapter;
 
 import android.view.LayoutInflater;
+import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
@@ -15,11 +16,12 @@ import java.util.List;
 
 public class ConfigAdapter extends RecyclerView.Adapter<ConfigAdapter.ViewHolder> {
 
-    private final OnClickListener mListener;
+    private final OnClickListener listener;
     private List<Config> mItems;
+    private boolean readOnly;
 
     public ConfigAdapter(OnClickListener listener) {
-        this.mListener = listener;
+        this.listener = listener;
     }
 
     public interface OnClickListener {
@@ -29,6 +31,11 @@ public class ConfigAdapter extends RecyclerView.Adapter<ConfigAdapter.ViewHolder
         void onDeleteClick(Config item);
     }
 
+    public ConfigAdapter readOnly(boolean readOnly) {
+        this.readOnly = readOnly;
+        return this;
+    }
+
     public ConfigAdapter addAll(int type) {
         mItems = Config.getAll(type);
         mItems.remove(type == 0 ? VodConfig.get().getConfig() : LiveConfig.get().getConfig());
@@ -36,9 +43,11 @@ public class ConfigAdapter extends RecyclerView.Adapter<ConfigAdapter.ViewHolder
     }
 
     public int remove(Config item) {
+        int position = mItems.indexOf(item);
+        if (position == -1) return -1;
         item.delete();
-        mItems.remove(item);
-        notifyDataSetChanged();
+        mItems.remove(position);
+        notifyItemRemoved(position);
         return getItemCount();
     }
 
@@ -57,11 +66,12 @@ public class ConfigAdapter extends RecyclerView.Adapter<ConfigAdapter.ViewHolder
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
         Config item = mItems.get(position);
         holder.binding.text.setText(item.getDesc());
-        holder.binding.text.setOnClickListener(v -> mListener.onTextClick(item));
-        holder.binding.delete.setOnClickListener(v -> mListener.onDeleteClick(item));
+        holder.binding.text.setOnClickListener(v -> listener.onTextClick(item));
+        holder.binding.delete.setVisibility(readOnly ? View.GONE : View.VISIBLE);
+        holder.binding.delete.setOnClickListener(v -> listener.onDeleteClick(item));
     }
 
-    static class ViewHolder extends RecyclerView.ViewHolder {
+    public class ViewHolder extends RecyclerView.ViewHolder {
 
         private final AdapterConfigBinding binding;
